Clean up flake and mac modules

+++ nixos_flake/modules/user/shell/work.nix
@@ -12,41 +12,39 @@ let
     with pkgs.google-cloud-sdk.components; [ gke-gcloud-auth-plugin ]
   );
 in
-{
-  home.packages =
-    (with pkgs; [
-      # TODO: may requires API token (disabled on our instance), but check other sign-in methods
-      # jira-cli-go
-      # BUG: doesn't get added to PATH and can't be executed
-      _1password # 1password cli, command "op"
+if envVars.system.isWork then
+  {
+    home.packages =
+      (with pkgs; [
+        # TODO: may requires API token (disabled on our instance), but check other sign-in methods
+        # jira-cli-go
+        # BUG: doesn't get added to PATH and can't be executed
+        _1password # 1password cli, command "op"
 
-      gh # github CLI
+        vault # takes forever to build; pls keep it pinned
+        openshift # os CLI
+        nodejs_22 # required for copilot-lua
+        ansible
+        sshpass
+      ])
+      ++ (with pkgs-unstable; [
+        vimPlugins.copilot-lua
+        # vimPlugins.CopilotChat-nvim
+      ])
+      ++ [ gdk-with-extras ];
 
-      vault # takes forever to build; pls keep it pinned
-      openshift # os CLI
-      nodejs_22 # required for copilot-lua
-      ansible
-      sshpass
-    ])
-    ++ (with pkgs-unstable; [
-      vimPlugins.copilot-lua
-      # vimPlugins.CopilotChat-nvim
-    ])
-    ++ [ gdk-with-extras ];
+    # create a symlink to work configs
+    home.activation.createWorkConfigs =
+      lib.hm.dag.entryAfter [ "writeBoundary" ] # sh
+        ''
+          rm -rf $HOME/.config/work
+          ln -sf ${envVars.NIX_DOTS}/work $HOME/.config/work
+        '';
 
-  # create a symlink to work configs
-  home.activation.createWorkConfigs =
-    lib.hm.dag.entryAfter [ "writeBoundary" ] # sh
-      ''
-        rm -rf $HOME/.config/work
-        ln -sf ${envVars.NIX_DOTS}/work $HOME/.config/work
-      '';
-
-  programs.zsh.sessionVariables = {
-    WORK_ENV = if envVars.system.isWork then "1" else "0";
-    HOME_ENV = if envVars.system.isWork then "0" else "1";
-  };
-  programs.zsh.shellAliases = {
-    ghc = "gh copilot";
-  };
-}
+    programs.zsh.sessionVariables = {
+      WORK_ENV = if envVars.system.isWork then "1" else "0";
+      HOME_ENV = if envVars.system.isWork then "0" else "1";
+    };
+  }
+else
+  { }
